% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NYT_Bestsellers.R
\name{NYT_BestSellers}
\alias{NYT_BestSellers}
\title{NYT Bestsellers}
\usage{
NYT_BestSellers(date = "current", category = "hardcover-fiction")
}
\arguments{
\item{date}{This is a date}

\item{category}{This is a qcategory}
}
\value{

}
\description{
NYT Bestsellers
}
\examples{
NYT_BestSellers(date = "2020-12-01")
}
